+++ randovania/games/cave_story/layout/preset_describer.py
@@ -33,7 +33,8 @@ class CSPresetDescriber(GamePresetDescriber):
             ],
             "Game Changes": [
                 message_for_required_mains(
-                    configuration.ammo_pickup_configuration, {"Missiles need main Launcher": "Missile Expansion"}
+                    configuration.ammo_pickup_configuration,
+                    [("Missiles need main Launcher", "Missile Expansion", False)],
                 ),
                 {"No falling blocks in B2": configuration.no_blocks},
             ],