+++ randovania/games/dread/layout/preset_describer.py
@@ -85,9 +85,10 @@ class DreadPresetDescriber(GamePresetDescriber):
             "Game Changes": [
                 message_for_required_mains(
                     configuration.ammo_pickup_configuration,
-                    {
-                        "Power Bomb needs Main": "Power Bomb Expansion",
-                    },
+                    [
+                        ("Power Bombs do not need Main", "Power Bomb Tank", True),
+                        ("Flash Shift does not need Main", "Flash Shift Upgrade", True),
+                    ],
                 ),
                 {
                     "Open Hanubia Shortcut": configuration.hanubia_shortcut_no_grapple,
@@ -102,7 +103,7 @@ class DreadPresetDescriber(GamePresetDescriber):
                     "X Starts Released": configuration.x_starts_released,
                 },
                 {
-                    "Power Bomb Limitations": configuration.nerf_power_bombs,
+                    "Disabled Power Bomb Limitations": not configuration.nerf_power_bombs,
                 },
             ],
             "Environmental Damage": _format_environmental_damage(configuration),